+++ src/main/java/com/example/gcj/dto/blog_reaction/BlogReactionResponseDTO.java
@@ -12,6 +12,6 @@ public class BlogReactionResponseDTO {
     private long id;
     private long blogId;
     private long userId;
-    private int reactionId;
-    private int rating;
+    private Long reactionId;
+    private Integer rating;
 }
